fix: map 검색기능 세분화
