+++ recycle/models.py
@@ -21,4 +21,3 @@ class Recycle(models.Model):
     
     def __str__(self):
         return self.name
-
