+++ templates/recycle/map.html
@@ -13,7 +13,9 @@
         <div class="option">
             <div>
                 <form onsubmit="searchPlaces(); return false;">
-                    키워드 : <input type="text" value="" id="keyword" size="15"> 
+                    찾고싶은 의류수거함이 속한 '동'을 검색해 주세요. 
+                    (현재 성북구 내 의류수거함만 지원하고 있습니다)
+                    ex. 돈암동, 돈암  <input type="text" value="" id="keyword" size="15"> 
                     <button type="submit">검색하기</button> 
                 </form>
             </div>
@@ -26,5 +28,5 @@
 {% endblock %}
 {% block script %}
     <script type="text/javascript" src="//dapi.kakao.com/v2/maps/sdk.js?appkey={{kakao_api_key}}&libraries=services"></script>
-    <script src='/static/js/recycle/temp.js'></script>
+    <script src="{% static 'js/recycle/map.js' %}"></script>
 {% endblock %}
